Make strlen be correct for strings bigger than int

+++ lib/my/src/my_strlen.c
@@ -9,7 +9,7 @@
 
 size_t my_strlen(char const *str)
 {
-    int result = 0;
+    size_t result = 0;
 
     while (*str++ != '\0')
         ++result;
